requirements.txt
falcon
fastapi
sqlalchemy
tortoise-orm
asyncpg
gunicorn
uvicorn
locust
psycopg2-binary
meinheld
gunicorn[gevent]
gino[starlette]

